requirements.txt
# Core dependencies
PyQt6
numpy
scipy
matplotlib
sounddevice
soundfile
pydub
audioop-lts

# Configuration
pydantic-settings
python-dotenv

# API
fastapi[standard]
python-multipart
httpx

# Auth & Payments
PyJWT
cryptography
stripe

# Testing
pytest
pytest-cov

# Optional: LLM providers (uncomment as needed)
# openai
# anthropic
# tavily-python
